JS: Added initializeImageArray()
This function is designed to keep an array that keeps track of the image
data. The data on the canvas is not enough information to be able to
sort the pixels back into the original image. Therefore this function is
required to attribute an index to each pixel and store the object
containing the color and index in an array.