JS: Massive Improvement to shuffleCanvas() Function
Previously the shuffleCanvas function created an array of indices,
selected two at random, swapped the data at those indices, then removed
the indices from the array so that they would not get selected again,
then repeated until the array of indices was empty.

This was needlessly resource demanding and overly complex.

The algorithm has been replaced to iterate through the array and select
a random index between the current index and the end of the array and
swapped the two, then repeated for the next index, until it iterated
through the full data array.