JS: Now only calls initializeImageArray() On Load
The initializeImage() should only be called when the original image is
loaded as intended. If it is called after the image is changed it will
sort to the state the image was in when the function was called.

Therefore it should not be called on button press, but on load.