HTML: LoadImage() is no longer called onload.
Previously this was called to load in the image and have it ready for
when the user wants it. Due to upcoming changes it will no longer be
preferable to default to loading one in until the user makes a
selection. There would be no benefit to call this onload as it will be
done again later.